feat: ok for test bitseed with playwright
