+++ sdk/bitseed-sdk/.gitignore
@@ -25,3 +25,7 @@ yarn-error.log
 /playwright-report/
 /blob-report/
 /playwright/.cache/
+/test-results/
+/playwright-report/
+/blob-report/
+/playwright/.cache/
