+++ sdk/bitseed-sdk/package.json
@@ -11,16 +11,22 @@
     "build": "tsup",
     "test": "npm run test:unit && npm run test:e2e",
     "test:unit": "jest src",
-    "test:e2e": "playwright test"
+    "test:e2e": "npm run build && playwright test -c playwright-ct.config.ts"
   },
   "devDependencies": {
+    "@playwright/experimental-ct-react": "^1.41.2",
     "@playwright/test": "^1.41.2",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.11.19",
+    "@types/react": "^18.2.55",
     "jest": "^29.7.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "ts-jest": "^29.1.2",
     "tsup": "^8.0.2",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "vite-plugin-node-polyfills": "^0.21.0",
+    "vite-plugin-wasm": "^3.3.0"
   },
   "dependencies": {
     "@sadoprotocol/ordit-sdk": "^2.3.12",
