+++ sdk/bitseed-sdk/playwright/index.tsx
@@ -1,0 +1,10 @@
+// Import styles, initialize component theme here.
+// import '../src/common.css';
+
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+ReactDOM.render(
+  <React.StrictMode></React.StrictMode>,
+  document.getElementById('root')
+);
