+++ sdk/bitseed-sdk/src/api/index.ts
@@ -1,2 +1,3 @@
 export * from './rooch-bitseed-api.interface'
 export * from './rooch-bitseed-api'
+export * from './rooch-bitseed-api.mock'
