--- sdk/bitseed-sdk/src/index.test.ts
@@ -1,3 +1,0 @@
-test('two plus two is four', () => {
-  expect(2 + 2).toBe(4);
-});
