+++ sdk/bitseed-sdk/src/interfaces/api.interface.ts
@@ -7,6 +7,7 @@ export interface DeployOptions {
 }
 
 export interface APIInterface {
+  name(): string;
   deploy(tick: string, max: number, generator: Generator, opts?: DeployOptions): Promise<InscriptionID>;
   mint(tick: string, amt: number, attributes?: Map<string, string>): Promise<InscriptionID>;
   merge(a: InscriptionID, b: InscriptionID): Promise<InscriptionID>;
