+++ sdk/bitseed-sdk/src/utils/index.ts
@@ -1,0 +1,3 @@
+export function add(a: number, b: number) {
+  return a + b
+}
